requirements.txt
DMXEnttecPro==0.4
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
pygame==2.6.1
Pygments==2.19.2
pyserial==3.5
pytest==9.0.2
python-osc==1.9.3
PyYAML==6.0.3
